+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/SystemTestDrivers.java
@@ -21,6 +21,7 @@ import sleeper.systemtest.dsl.compaction.CompactionDriver;
 import sleeper.systemtest.dsl.gc.GarbageCollectionDriver;
 import sleeper.systemtest.dsl.ingest.DirectBulkImportDriver;
 import sleeper.systemtest.dsl.ingest.DirectIngestDriver;
+import sleeper.systemtest.dsl.ingest.EksBulkImportDriver;
 import sleeper.systemtest.dsl.ingest.IngestBatcherDriver;
 import sleeper.systemtest.dsl.ingest.IngestByAnyQueueDriver;
 import sleeper.systemtest.dsl.ingest.IngestByQueue;
@@ -47,7 +48,6 @@ import sleeper.systemtest.dsl.sourcedata.DataGenerationTasksDriver;
 import sleeper.systemtest.dsl.sourcedata.GeneratedIngestSourceFilesDriver;
 import sleeper.systemtest.dsl.sourcedata.IngestLocalFilesDriver;
 import sleeper.systemtest.dsl.sourcedata.IngestSourceFilesDriver;
-import sleeper.systemtest.dsl.statemachine.StateMachineDriver;
 import sleeper.systemtest.dsl.statestore.StateStoreCommitterDriver;
 import sleeper.systemtest.dsl.statestore.StateStoreCommitterLogsDriver;
 import sleeper.systemtest.dsl.util.PollWithRetriesDriver;
@@ -87,7 +87,7 @@ public interface SystemTestDrivers {
 
     DirectBulkImportDriver directEmrServerless(SystemTestContext context);
 
-    StateMachineDriver stateMachine(SystemTestContext context);
+    EksBulkImportDriver stateMachine(SystemTestContext context);
 
     IngestBatcherDriver ingestBatcher(SystemTestContext context);
