+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/ingest/EksBulkImportDriver.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.dsl.statemachine;
+package sleeper.systemtest.dsl.ingest;
 
 import java.util.List;
 import java.util.Map;
 
 /**
- * Queries the execution status of bulk import jobs in an AWS Step Functions state machine.
+ * Runs checks on bulk import jobs that are specific to running in EKS.
  */
-public interface StateMachineDriver {
+public interface EksBulkImportDriver {
 
     /**
      * Retrieves the execution status of each bulk import job in the state machine.
@@ -30,7 +30,7 @@ public interface StateMachineDriver {
      * @param  stateMachineArn the ARN of the Step Functions state machine
      * @param  tableId         the table ID used to derive execution names
      * @param  jobIds          the bulk import job IDs to query
-     * @return a map of job ID to execution status (e.g. "SUCCEEDED", "FAILED", "RUNNING")
+     * @return                 a map of job ID to execution status (e.g. "SUCCEEDED", "FAILED", "RUNNING")
      */
     Map<String, String> getJobExecutionStatuses(String stateMachineArn, String tableId, List<String> jobIds);
 }
